backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.20
pypdf==5.9.0
chromadb==1.0.15
google-genai==1.36.0
pytest==8.3.4
